backend/requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.3.4
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
black==21.6b0
bleach==3.3.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
decorator==5.0.9
defusedxml==0.7.1
diff-match-patch==20200713
dj-rest-auth==2.1.7
Django==3.2.4
django-admin-honeypot==1.1.0
django-allauth==0.44.0
django-appconf==1.0.4
django-cities==0.6
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-extensions==3.1.3
django-filter==2.4.0
django-imagekit==4.0.2
django-import-export==2.5.0
django-model-utils==4.1.1
django-request==1.5.6
django-simple-history==3.0.0
django-sql-explorer==2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
drf-extra-fields==3.1.1
drf-yasg==1.20.0
entrypoints==0.3
et-xmlfile==1.1.0
idna==2.10
inflection==0.5.1
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
Markdown==3.3.4
MarkupPy==1.14
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.7
packaging==20.9
pandocfilters==1.4.3
parso==0.8.2
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==8.2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.1.0
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.9
ruamel.yaml.clib==0.2.2
Send2Trash==1.5.0
six==1.16.0
sqlparse==0.4.1
swagger-spec-validator==2.7.3
swapper==1.1.2.post1
tablib==3.0.0
terminado==0.10.1
testpath==0.5.0
toml==0.10.2
tornado==6.1
tqdm==4.61.1
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.5
wcwidth==0.2.5
webencodings==0.5.1
xlrd==2.0.1
XlsxWriter==1.4.3
xlwt==1.3.0
